sfx system revamp + docmentation

+++ release.py
@@ -1,15 +1,22 @@
 import os
 import shutil
 import zipfile
+import argparse
 
 import PyInstaller.__main__
+
+# argument parser
+parser = argparse.ArgumentParser(description='Package application for release.')
+parser.add_argument('-d', '--destination', type=str, required=True, help='Destination directory for the release')
+parser.add_argument('-z', '--zip', action='store_true', help='Whether to zip the output (default: False)')
+args = parser.parse_args()
 
 # ========== #
 # PARAMETERS #
 # ========== #
 
-RELEASE_DESTINATION = './releases'
-RELEASE_NAME = 'linux-v0a-04-28'
+RELEASE_DESTINATION = args.destination
+ZIP_OUTPUT = args.zip
 
 # setup
 os.makedirs( './.release/static/css', exist_ok=True )
@@ -43,7 +50,7 @@ shutil.copy( './development/static/css/styles.css', './.release/static/css' )
 
 # js
 cjs = ''
-scripts = [ './development/static/js/soundboard.js', './development/static/js/menu-bar.js', './development/static/js/details.js', './development/static/js/util.js', './development/static/js/edit.js' ]
+scripts = [ './development/static/js/soundboard.js', './development/static/js/soundeffect.js', './development/static/js/menu-bar.js', './development/static/js/details.js', './development/static/js/util.js', './development/static/js/edit.js' ]
 
 for script in scripts:
     with open( script, 'r' ) as f:
@@ -80,15 +87,22 @@ shutil.copy( './development/tracks/.gitkeep', './Dungeon-Lute/tracks' )
 shutil.copy( './development/sfx/.gitkeep', './Dungeon-Lute/sfx' )
 
 shutil.copy( './assets/readme.txt', './Dungeon-Lute' )
+shutil.copy( './LICENSE', './Dungeon-Lute')
 
-# zip  release
-def zip_folder( folder_path, zip_name ):
-    with zipfile.ZipFile( zip_name, 'w', zipfile.ZIP_DEFLATED ) as zipf:
-        for root, dirs, files in os.walk( folder_path ):
-            for file in files:
-                file_path = os.path.join( root, file )
-                zipf.write( file_path, os.path.relpath( file_path, os.path.join( folder_path, '..' ) ) )
+if ZIP_OUTPUT:
 
-zip_folder( './Dungeon-Lute', RELEASE_NAME )
-shutil.move( f'./{RELEASE_NAME}', RELEASE_DESTINATION )
-shutil.rmtree( './Dungeon-Lute' )
+    # zip  release
+    def zip_folder( folder_path, zip_name ):
+        with zipfile.ZipFile( zip_name, 'w', zipfile.ZIP_DEFLATED ) as zipf:
+            for root, dirs, files in os.walk( folder_path ):
+                for file in files:
+                    file_path = os.path.join( root, file )
+                    zipf.write( file_path, os.path.relpath( file_path, os.path.join( folder_path, '..' ) ) )
+
+    zip_folder( './Dungeon-Lute', f'./zip' )
+    shutil.move( f'./zip', f'{RELEASE_DESTINATION}' )
+    shutil.rmtree( './Dungeon-Lute' )
+    
+else:
+    os.rename( './Dungeon-Lute', f'./prgm' )
+    shutil.move( f'./prgm', f'{RELEASE_DESTINATION}' )
